Builds at least

+++ src/engine/directions.h
@@ -46,7 +46,7 @@ enum class TDirection : uint8_t {
     ALL_DIRECTIONS = ALL_CARDINAL_DIRECTIONS | ALL_DIAGONAL_DIRECTIONS
 };
 
-inline const std::unordered_map<TDirection, glm::vec2> isometric_direction_vectors { {
+inline const std::unordered_map<TDirection, glm::ivec2> isometric_direction_vectors { {
     { TDirection::NORTH, { 2, -1 } }, //
     { TDirection::NORTH_WEST, { 0, -1 } }, //
     { TDirection::WEST, { -2, -1 } }, //
@@ -65,7 +65,7 @@ inline const std::unordered_map<TDirection, glm::vec2> isometric_direction_vecto
     // { TDirection::NORTH_EAST, { 0, -0 } } //
 } };
 
-inline const std::unordered_map<TDirection, glm::vec2> direction_vectors { {
+inline const std::unordered_map<TDirection, glm::ivec2> direction_vectors { {
     { TDirection::NORTH, { 0, -1 } }, //
     { TDirection::NORTH_WEST, { -1, -1 } }, //
     { TDirection::WEST, { -1, 0 } }, //
@@ -76,7 +76,7 @@ inline const std::unordered_map<TDirection, glm::vec2> direction_vectors { {
     { TDirection::NORTH_EAST, { 1, -1 } } //
 } };
 
-inline const std::unordered_map<glm::vec2, TDirection> vector_directions { {
+inline const std::unordered_map<glm::ivec2, TDirection> vector_directions { {
     { { 0, -1 }, TDirection::NORTH }, //
     { { -1, -1 }, TDirection::NORTH_WEST }, //
     { { -1, 0 }, TDirection::WEST }, //
